+++ hippylib/utils/__init__.py
@@ -15,8 +15,9 @@
 
 from .checkDolfinVersion import dlversion, checkdlversion
 from .deprecate import deprecated
+from .experimental import experimental
 from .vector2function import vector2Function
 from .random import Random, parRandom
 from .parameterList import ParameterList
-from .warnings import hIPPYlibDeprecationWarning
+from .warnings import hIPPYlibDeprecationWarning, hIPPYlibExperimentalWarning
 from . import nb