+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Development version
   and from related classes.
 - Use `argparse` to set parameters in application drivers from command line
 - Use `dl.XDMFFile` to export solutions for visualization in Paraview in all application drivers
+- Implement accuracy enhanced SVD algorithm in `randomizedSVD.py`
+- Increase coverage of unit testing in CI
 
 
 Version 2.3.0, released on Sept 6, 2019